Just print the differences between messages_{LOCALE} properties and print a warning message instead of build failure.
